+++ djangae/db/utils.py
@@ -16,7 +16,7 @@ from google.appengine.api.datastore import Key, Query
 
 #DJANGAE
 from djangae.utils import memoized
-from djangae.indexing import special_indexes_for_column, REQUIRES_SPECIAL_INDEXES
+from djangae.db.backends.appengine.indexing import special_indexes_for_column, REQUIRES_SPECIAL_INDEXES
 from djangae.db.backends.appengine.dbapi import CouldBeSupportedError
 from djangae.db.backends.appengine import POLYMODEL_CLASS_ATTRIBUTE
